requirements.txt
# =============================================================================
# VoiceClone - Requirements File
# Python 3.10+ required
# =============================================================================

# =============================================================================
# Core System Dependencies
# =============================================================================
runpod>=1.6.0
psutil>=5.9.0

# =============================================================================
# PyTorch Stack (CUDA 12.1+ support)
# For CPU-only: pip install torch torchvision torchaudio
# For CUDA: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
# =============================================================================
torch>=2.5.0
torchaudio>=2.5.0
torchvision>=0.20.0

# =============================================================================
# Chatterbox TTS Core Dependencies
# =============================================================================
numpy>=1.26.0
librosa==0.11.0
soundfile>=0.12.0
s3tokenizer
transformers==4.46.3
diffusers==0.29.0
resemble-perth==1.0.1
perth>=1.0.0
conformer==0.3.2
safetensors>=0.4.0
huggingface-hub>=0.20.0
accelerate>=0.20.0

# =============================================================================
# Story Narrator & LLM Dependencies
# =============================================================================
# Google Gemini (default LLM)
google-generativeai>=0.8.0

# Optional: Alternative LLM providers (uncomment if needed)
# openai>=1.0.0
# anthropic>=0.18.0

# =============================================================================
# Environment & Configuration
# =============================================================================
python-dotenv>=1.0.0
pydantic>=2.0.0
python-json-logger>=2.0.0

# =============================================================================
# Database
# =============================================================================
psycopg2-binary>=2.9.9

# =============================================================================
# Authentication & Security
# =============================================================================
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
python-multipart>=0.0.6

# =============================================================================
# Web Interface
# =============================================================================
gradio>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# =============================================================================
# Monitoring & Metrics (Optional)
# =============================================================================
prometheus-client>=0.19.0
tqdm>=4.66.0

# =============================================================================
# Development & Testing
# =============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# =============================================================================
# Additional Utilities
# =============================================================================
requests>=2.31.0
pyyaml>=6.0
